be less buggy when adding rows. Refs UICIRC-29, UICIRC-30.
When adding a new row with `fields.unshift()` instead of
`fields.push()`, the new row is a clone of the topmost row instead of an
empty row. Huh? See UICIRC-29.

Adding new rows as part of the FieldArray component if the `fields`
pseudo-array is empty causes a console error because those pseudo-array
manipulators (push, pop, shift, unshift) aren't actually mutators, the
issue state changes to the store. So there needs to be an alternative
implementation to achieve that effect. See UICIRC-30.