update version in order to depend upon this change
Tests in other repositories may twiddle the checkboxes on the Settings >
Circulation > Scan ID page in order to, e.g., enter a user by barcode
rather than by username (the default). In order to do so, they need
access to the save-button in order to save the change after ticking the
checkbox.

Refs UITEST-20.