always show checkout-scanning checkboxes
Sometimes, the checkboxes for checkout-scanning would not show. I don't
know why, exactly, but it's clear that was not the desired behavior, and
it's clear that when they didn't show, then we couldn't check those
boxes as we wanted to in integration tests, and then those tests would
fail. May the fail no longer.

Refs UICIRC-75