be lenient with react deps
Workspace builds are all kinds of whacked. It appears that, maybe, we
are getting conflicting versions of react in the bundle because of ~
deps on a version that is older than the one provided by ^ deps from
other modules, maybe? It's not totally clear. But using a ^ dep here
will at least be consistent with other apps. Consistency is good.