correctly format the dataOptions for a Select
`<Select>` accepts a `dataOptions` prop that expects to be an array of
key-value pairs with the props `value` and `label`. It was receiving,
instead, an array of `<option>` elements which caused rendered list to
show up as a series of empty `<option />`s. TBH, I'm not sure how this
ever worked.

Refs STRIPES-602