UITEN-35 restore settings menu
When `settings.circulation.enabiled` was removed in #360 under the auspices of
UITEN-35, the permission sets that relied on it should have been updated
to include `settings.enabled` in order to cause Settings to be
accessible.

Removing `settings.circulation.enabled` _was_ the right thing to do; it was
functionally a do-nothing permission because the individual menu items
are controlled by more granular permissions. BUT we do need to make sure
that the top-level Settings menu is displayed when any of those
granular permissions are granted.

Refs UITEN-35