UIIN-678 avoid buggy react-to-print version (#394)
`react-to-print` v1 depends directly on React with a caret dep on
version 16. This has the effect of pulling in the most-recent React 16.x
build into any bundle that uses this module, even if that module has a
more strict tilde dep on a less-than-current React minor release.
Whatnow? Say an app depends on React ~16.8.0; React 16.9.0 still gets
into the bundle because of the loose (caret) direct dependency in
react-to-print v1. This was fixed in
https://github.com/gregnb/react-to-print/pull/53 and appears to be the
straw that broke the v1.0-camel's back.

v2 also changed from using a `window` to an `iframe` to handle printing,
but does not automatically clean up the iframe after printing is
finished. This appears to be an oversight, but rather than introduce
another breaking change, they just added the prop in release 2.3.0.

Refs UIIN-678