UICIRC-430 improve location display in circ rules
The CSS rules governing the line-height of a single location-item are
completely impenetrable to me, but removing the `height` attribute on
the container element somehow, mysteriously, fixes the display. It
appears that what's happening is a bunch of block-level elements with a
too-short line-height are stacked, such that the line below clips that
above, but I couldn't figure out how to change the line-height for a
single element anywhere.

Refs UICIRC-430