[UICIRC-1238] Unconditional hook, fixes slip-editor HTML mode
Invocation of `useFormState` in `StaffSlipTemplateContentSection.js`
was conditional on the existence of the `useFormState` function (which
was not present when running tests). Probably as a result of this,
switching between HTML and WYSIWYG editors for a staff slip did not
work in minified builds such as that of snapshot.

The invocation is no longer conditional, and I have provided a mock so
that the test works.